Handle 32-bit sample size (just in case)

+++ Audio/Decoders/FLACStreamDecoder.m
@@ -45,7 +45,7 @@ writeCallback(const FLAC__StreamDecoder *decoder, const FLAC__Frame *frame, cons
 	}
 	
 	float	*floatBuffer	= (float *)writePointer;
-	float	scaleFactor		= (1 << frame->header.bits_per_sample);
+	double	scaleFactor		= (1LL << frame->header.bits_per_sample);
 	
 	for(sample = 0; sample < frame->header.blocksize; ++sample) {
 		for(channel = 0; channel < frame->header.channels; ++channel) {
